Change retry delay from 1ms to 0 in ZooKeeperSeedNodeProvider.
Motivation:

Retry delay of 1ms was initially set as it wasn't clear if zero value
(i.e. no wait) is supported. Now it is clear that we can use no-wait
policy and hence should set it to 0.

Modification:
Change retry delay from 1ms to 0.

Results:
No useless delay and cleaner code in ZooKeeperSeedNodeProvider.