Confirmations for send/broadcast operations in the MessagingService.
Motivations:

When a message is submitted to a remote node via send()/broadcast()
methods there is no way to know on whether the operation was successful
or failed with an error on a remote peer.

As a workaround it is possible to use request()/aggregate() methods that
would return an empty response, however it would be much better if we
could provide such functionality out of the box and allow users to
switch between confirmation and non-confirmation modes for send() and
broadcast() operations.

Modifications:

 - Add MessagingChannel#withConfirmReceive(boolean) method that would
   allow users to enable/disable confirmations.
 - Extend messaging protocol so that it would send an empty response
   message when a message gets processed by the channel's receiver.
 - If confirmations are enabled then notify callback/future objects
   a send()/broadcast() operations only when confirmation is received
   from a remote peer.
 - Make confirmations enabled by default

Results:

MessagingChannel's send()/broadcast() operations can provide a feedback
on whether the operation was successful or failed with an error on
a remote peer.